backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
SpeechRecognition==3.8.1
python-dotenv==0.19.0
openai==1.8.0
Werkzeug==2.0.1
ffmpeg-python>=0.2.0
numpy>=1.21.0
httpx==0.24.0
